requirements.txt
Flask
python-dotenv
pygame
Pillow
tqdm
watchdog